Bump libMTP dependency to version 1.0.0.
I think it really makes sense to increase this, as the version we dependeded on
before (0.3.0) is hopelessly outdated, and buggy. As far as I know, all
distros are able to provide version 1.0.0, but we could notify the packagers
to make sure. Also, version 1.0.0 is API compatible.

CCMAIL: amarok-devel@kde.org