Fix a crash in ScanResultProcessor's destructor.
Despite all the extra checks I put for null pointers in the destructor,
this wasn't actually the problem. The problem, which could occur in a
few ways, was that in certain instances values were either not removed
from a hash when they should have been, or added to a hash when they
shouldn't have been. Then when the destructor ran, it would hit two
values pointing to the same QStringList, and the second access would
cause a crash.

Thanks to Mikko for doing some heavy legwork helping me narrow down the
cause, and Tm_T and hyper_ch for their help.

BUG: 220532