Ensure that there exists a collection path in the local collection that has enough free space to copy the intended tracks to and still have at least 95% of the drive free, with a minimum of 5 megs, and, of course, be writable. All collection folders that do not satisfy these conditions are filtered out of the collection folder drop down in the organize dialog, so only folders that have enough space can be chosen.
If there is not enough free space the organize dialog is not even shown, the
entire process is aborted with a statusbar()->longmessage explaining the issue.

Also, isWriteable() has been modified to take into account the above conditions.
i.e., if >= 95% of the each path is full the collection will return that it is
not writeable.

BUG: 126283