first run: handle QDesktopServices::MusicLocation more graciously
On first run, a configuration dialog appears for the user to select
collection folders. However, if MusicLocation is set, it is scanned
automatically. This is unintuitive since scanning should only begin
after the user has chosen any folders.

This patch checks MusicLocation on first run, and if it is available a
message box is shown asking if the user wants to use that as a
collection folder.

If "Yes", the dialog returns to the main window and that path is scanned
into the collection. If "No", Amarok's configuration dialog pops up and
the user selects the folders manually (the previous behaviour). If "Cancel"
then nothing happens.