Patch for bug 217159: Enter key in playlist search does nothing
Completes the behavior requested in this wish by clearing the search term.
Clearing the search term automatically when they finish a search makes for a better user experience because the user would normally erase the old term when they were looking for the next thing they want to search for.