Add "--savelocation" to scanner help output.
Apparently this is a new feature of the scanner that is used by Amarok. To
make easier for the user to tell if the scanner version is correct,
I have added it to --help. @Jeff: Please check if the text is OK.

Also see: http://forum.kde.org/viewtopic.php?f=115&t=84293

CCMAIL: mitchell@kde.org