Add check for correct collection scanner version.
If the version is wrong, Amarok now reports this with an error dialog.
The check _almost_ works, but gets wrong results, although the version
is correct (wrong endline truncating or so). So I've disabled it for now.
Will fix up later (or someone else fix it, if you can spot the bug).