Add flac and ogg to the list of allowed format when copying to a UMS device.
This stuff needs a real config dialog, but this change at least makes amarok usefull with the my N900.
Also TODO: the error message when tracks cannot be copied, for any reason, always defaults to "tracks are already on device" even if the reason is another one, such as the format not being supported by the device.
Ideally it should show a dialog with the failure reason for wach track that is not copied, and it seems the basis of this is already in the code, just not used yet.
~lart xevix