Playlist performance: make GroupingProxy just-in-time instead of pre-calculated.
The old GroupingProxy code did an O(n) scan over the whole playlist every time
SortProxy emits a Qt 'layoutChanged' signal. But that can happen at every "Play
next track" event, if QSortFilterProxyModel is feeling trigger-happy about
'dataChanged' signals...

There is still some linear scanning going on in this new code. But it is
O(group size), and very large group sizes are too unlikely to worry about.

This commit is part of the fix for Bug 210345. (slow "play next track" when the
playlist is large)