Don't generate and cache a pixmap for each slider position/size, but paint the mid part tiling.
This saves a lot of memory* and some (pot. many) cpu cycles.
The current update clipping doesn't allow stretching sliders anyway, also this
is not required by the current slider look.

*the xrestop thing:
a 32bit pixmap takes w*h*4 bytes
assuming the slider is 1000px wide and each size has it's own pixmap (16.6 minutes, or just several tracks)
we'll generate 1000 pixmaps with a mean width of ~500px and ~12px heigh (dep. on the UI style)
that's nearby 23MB just for the pixmap data or - if we overflow the cache, caching becomes superflous
(because the entry is discarded before it's queried next time)

If we ever need truely stretchable sliders or similar, please call on ;-)