Playlist FilterProxy: remove re-declaration of inherited signals.
In this case, there should not be any change in behaviour.

But re-declaring inherited signals is unnecessary and has potential for (very
hard to find) bugs in some situations.

See the recent 'GroupingProxy was being evil' commit for a long story.

Short story: for signals, we neither need nor want 2 vtable entries that have
the same name.