Playlist SortAlgorithms: make the "random" sort order persistent.
The existing SortProxy tries to solve the "random order shouldn't change at
every re-sort" problem by suppressing re-sorts as much as possible.
(By not letting QSortFilterProxyModel do normal handling of 'dataChanged'
signals).

That approach is iffy anyway, but it also causes a performance problem.

So let's fix the random sort problem in a more fundamental way, and that will
clear the way to improve SortProxy in an upcoming commit.