Playlist layout editor: it should not be possible to rename one of the default layouts
Currently, the logic that was supposed to catch this sort of change, was
wrong as it wasn't checking whether the layout being renamed is a
default one but whether the modified name belongs to a default layout
(yet, the error message was expressing the former).

The right thing to do is to disable the Rename action for default
layouts just like the Delete action is.