Revert "Add embedded cover art support"
As discussed on the mailing list, there are multiple problems with this patch. Besides the several technical issues, which Dan, Bart, and Jeff all bring up, there is the fact that we have been in feature freeze for 2 weeks, and this does not belong right now.

Andreas, we still want this feature, and we will be merging it back in as soon as 2.3 final is tagged. However, you can see the multiple comments that people have had, and things that we would like to be changed (e.g. no taglib code in sqlmeta).

Anyway, regardless of the technical quality, this still breaks our feature freeze for the 2.3.0 release, and given the ease with which we can develop in parallel in git, we have no reason to violate our freeze for this.

This reverts commit a0c4a6c706d6f4f6cf22740bfa4aec768a7c4b87.

Revert "One more for you, good friend ChangeLog."

This reverts commit 3be4fe9b8f4adca3d2f67db14b1f55240e679c80.

CCMAIL: amarok-devel@kde.org
CCMAIL: ecroy@gmx.net