Awful hack to "fix" BUG 229226 and 222129. Just a stopgap because of the v2.3 freeze.
This "fix" can have awful performance, but a proper fix needs to do some
invasive refactoring which is not appropriate during a freeze.

A delay before the next track is better than no next track at all.

This commit also makes no attempt to keep 'likelyNextTrack()' in sync, or to
fix any of the other track navigators. Classify as "very temporary hack".