fix crashes when deleting songs from an ipod
The ipod media device plugin's internal list of files was not correctly updated
when tracks have been copied to an ipod or when orphaned tracks have been added.
This resulted in songs being added to the ipod's database multiple times. Fix that
and try to work around problems resulting from the invalid databases containing
several entries for the same file.

BUG: 213097