cover fetcher: remove pending manual jobs if cover found dialog is closed
When entering a custom query for covers in the cover found dialog, the user
may cancel or close the dialog before all the results are fetched. This
causes the dialog to reappear with the newly fetched covers, which is
annoying.

Furthermore, covers retrieved from custom queries aren't associated with
any album for these newly created dialogs (because such queries should only
be available when a dialog has already been created from an album). So
crashing is ensured if a cover is chosen from this new dialog. Making sure
the dialog does not reappear solves this.

BUG: 230215