cover fetcher: associate each cover with metadata from fetched source
This commit paves the way for showing metadata of each item in the cover
found dialog. Previously there's album and artist name in the "details"
widget from the dialog; but they were static, dependent on whatever album
the dialog was created from. Anyway the details widget is planned to be
replaced with a much better side info pane.