Use album names instead of album pointers to group tracks into albums in the RepeatAlbumNavigator. Grouping but album pointers makes the navigator misbehave for meta types that use private album pointers.
An alternative would be to use the albums == operator, as this has been overridden to provide the correct behaviour for meta types where grouping has to be done by name. Yet another alternative would be to simply use teh grouping
info from the grouping proxy as that has this stuff well under control.