Always navigate browser with single-click
Previously, the browser used the system-wide setting for "activate", be
it single-click or double-click (this can be set in the KDE control
panel under "mouse settings").  In the case of the browser, requiring
double-click is unintuitive; the browser items look more like buttons,
which always respond to single clicks.

This changes the browser widget to always use single-click, no matter
what the system-wide setting for "activate" is.