Add a versioning system for Context View applets.
The purpose of this is the same as with our Plugins: Making sure that
deprecated applets from old Amarok installs no longer show up.

TODO: Discuss with Leo Franchi how to handle the problem with the config
loading. Currently if an applet is not loaded, a ghost of the applet appears.

TODO: Extend our bump_plugin_version.rb script to handle applet versioning.

CCMAIL: amarok-devel@kde.org