meta: album: use image location as primary key for bordered pixmap cache
BUG: 170146

The name of the album can be ambiguous, especially for compilations such
as "Greatest Hits", "Best of" etc. This causes the bordered pixmap cache
to store only the first album's cover, and subsequent albums with the
same name trying to retrieve their covers from the cache will use that
for display. The name is now a fallback if location is empty.