Make it possible to do Dolphin style editing of the paths inthe breadcrumb widget.
This works everywhere in amarok, so it is possible to navigate both in amarok and in the file system (entering the
path of a directory will show the file browser and have it navigate to the directory no matter what category is
currently showing). It is also possible to paste in any AmarokUrl and it will run.