This moves the items that were in meta/ (including meta itself) into core/
It also removes various references to meta/ and its subdirectories from CMakeLists. Please don't start adding these back. With the various meta/ directories around, in addition to some capabilities/ directories and the like, requiring you to
use the full path not only keeps you mindful of the strucure (if you add or change it around) but also ensures that you're referencing the directory and files that you think that you are.

Plus, this consistency will make it far easier for any further restructuring to be done, which is not unlikely.

CCMAIL:amarok-devel@kde.org

Interim step towards core/

Finish the move to core/