Playlist cleanup: make a clearer distinction between QAbstractItemModel and Playlist::AbstractModel.
Playlist::AbstractModel pretends it is a QAbstractItemModel by offering member
functions with identical names as QAbstractItemModel, but it doesn't actually
inherit from QAbstractItemModel.
(because Qt has very bad support for multiple inheritance)

That is a dirty hack; the existing code only gets away with it because it tries
to lexically override each member of QAbstractItemModel.

This is a first step towards cleaning this up.