The point of these confirmation dialogs is to show the user _exactly_ what operation is about to be performed. Merely showing the Artist+Title is not near enough information.
Undo Mark's change and use prettyUrl() for user presentation, this is
after all what it is for. This means collections need to correctly
implement prettyUrl() if their playableUrl() does not suffice.

I took the initiative to edit the MediaDevice implementation,
but I cannot test it.

CCMAIL:kretschmann@kde.org
CCMAIL:amarok-devel@kde.org