Meta->Playlists
Just to mess with everybody's heads: a rebase of a rebase commit message!

This is a combination of 2 commits.
The first commit's message is:

This is a combination of 2 commits.
The first commit's message is:

First part of Meta->Playlists

This is the 2nd commit message:

Second part of Meta->Playlists

This is the 2nd commit message:

Fix brokenness after rebase