Hackaround (not even worthy of being called a workaround) for the hang in the settings dialog when opening a service config panel.
This apparently happens whenever the vertical scrollbar has to be shown in the KPluginSelector, so perhaps this is really a bug there?
The Hackaround was to make the general config panel slightly taller, so that with the current selection of plugins the scrollbar is not needed. Of course, this is very fragile and can break on differnt styles, font sizes, more
services getting added, your cat having a bad hair day, ....