volume: Listen for signals from Phonon regarding volume/mute changes.
If a Phonon mixer (or PulseAudio mixer for Phonon with PulseAudio support)
changes the volume of the Amarok stream, we want the Amarok GUI
to reflect this change.

This commit does however introduce a feedback loop that will need to be fixed
to prevent the notification of a change ultimately triggering a change of
volume. Comments in the code highlight this problem.