volume: Do not set the volume to 0 when muted
This behaviour relates to amarok 1.x days where there was no independent
mute support. While it is cosmetically nice to not show the slider's
current value when muted, the quirks this introduces (e.g. when
unmuting automatically via scrolling the mouse wheel up or down)
are rather complicated to handle. If some kind of UI change of the slider
is desired then changing the colour would likely be more appropriate
than setting the value to 0.

This also helps with support when volume changes are pushed up
to amarok from lower down the stack (e.g. from Phonon)