EngineController: Add a position guesstimator for bounded playback.
I've noticed that for some backends, the position doesn't update as often as the
tickInterval.  To remedy this, we count the number of repeated ticks to
calculate a better position.

The motivation for this fix was for cue tracks that played a little bit over the
end in certain formats.

I would like to know if this is required, or perhaps the backend should handle
this and this is just a dirty hack.