APG: make PresetEditDialog modal
Workaround for a stupid Qt bug.  When you share a model across multiple
views in multiple dialogs, the views in other dialogs don't update
properly when the model is changed, even when properly using
removeRows() and insertRows().  Since there's really no reason for the
APG to open multiple copies of the PresetEditDialog for one preset, the
easiest workaround is just to make the Dialog modal.