XSPF playlist now saves queue.
Since Amarok currently saves the default playlist as XSPF, it now supports saving the queue
too. The XSPF specification allows arbitrary XML in the <extension/> tag. The application
has been identified with Amarok's home page so that other applications can ignore it.
Track URLs (<track>) are written as children of the <queue> element within the extension.