Use QStyle primitives to draw the expanders.
The green expanders on the root items in CB, User Playlists and Podcasts were a quick hack. I'm attemting to use the active QStyle now.

First attempt: using PE_IndicatorBranch
Another option is to use PE_IndicatorArrowRight and PE_IndicatorArrowDown (when expanded) though this is not sure to look good with alternative styles.

If this is OK I'll port this indicator style to the CollectionTreeItemDelegate.

CCMAIL:amarok-devel@kde.org