Remove extra delays of single click operations in CollectionTreeView
If KDE is running in single click mode, a noteable delay occurs when
e.g. expanding an item, starting drag for the PUD. Under double click
mode, there's no such delay. Further, this delay is set to
DoubleClickInterval, which makes the ui seem unresponsive.

This patch removes that delay and fixes some issues related to selection
of items.

BUG:222760
CCBUG:226533