Show the interactive cover dialog as soon as possible
Previously it is opened only after the first image was fetched. Also that
first image was fetched from Last.fm regardless of the chosen source
(except for Discogs). This is fixed as well.

The fetching speed has not changed, but I suspect it will *feel* quicker
since the dialog pops up faster :p .