New UpnpCollection
Separated the factory into separate file.
The collection no longer relies on HUpnp directly.
It uses the network:/ slave to monitor changes and
will use the upnp-ms:/ slave to provide the music.

Current status is that it adds new devices on the fly by monitoring the network using KDirLister.

Connections are not removed yet