Tweak the way the Google Mock cmake module works to make it work on my system.
With these changes the -L flag as gotten form gmock-config will be set
to GOOGLEMOCK_LIBRARIES, as well as the standard -lgmock -lgtest. Before
it was just doing -lgmock -lgtest, and ld couldn't find the location of
the libraries. This shouldn't break it for anyone, but who knows..
please test.

CCMAIL: amarok-devel@kde.org