removed checking for "large" in filename because this does not mean that a file is a cover thinking of "Back (large).jpg" at the end, the function also checks the filesize, what is more meaningfull.
the file what the last check is about is called "folder.jpg". this is from windows
world and has the effect, that in explorer view the folder get that image as icon.
people want to see the album cover in that place, thats why they name it that way.
there can also be a Image that is called "folder-back.jpg", so that check needs to be
changed to "compare" rather than "contains".

also added a quick check at start to quickly return if there is only one image to choose
the best.