Fixed track listing issue.
The bug was that each track was creating a new artist and overwriting the existing artist. This fixes it by holding all the artists/albums in member maps and finally setting them when all tracks are done.

Ideally I would like to directly use the MemoryCollection map since it seems unwieldy to have two declarations.

Thanks to nhn for suggesting what was wrong