Converted UpnpCollectionFactory to solely use Cagibi, which means we get direct access to the UDN.
The factory maintains a hash of active collections and removes collections when devices go offline.
The removal part should work but its not confirmed since the kio slave crashes (assertion error) when the device goes offline.

The friendly name is not resolved right now, but that should be pretty straightforward using deviceDetails