Prevent thumbnail cache pruning at every startup / settings change
When settings changed, App::applySettings() is called (fair enough); it is
also called every time Amarok starts up. In it, "unneeded" cover thumbnails
are deleted from the cache directory. Furthermore, an image is deemed
"unneeded" if its size is not from a list of hard coded sizes.

A better place to prune the cache would be in the cover manager, and
initiated by the user. To be done later.

CCBUG:240300