FastForwardImporter: rework, replace obsolete methods, check whether imported files belong to collection
this change splits the code to several smaller chunks and adds new logic:
when tracks are imported, it is determined which collection location the
track should belong to, it's collection location is recorded and then
the track is added in the correct collection location. Tracks not belonging
to any collection location are discarded.

This way we avoid deprecated primaryCollection() call.

Should fix a memleak where we leaked one StatisticsCapability object per imported
track.

Fixes b.k.o bug 218999