Implemented removeTrack and the ability to invalidate tracks based on location.
The m_tracksInContainer hash maps a upnp-ms:// path of the directory to
which the musicTrack item belongs. createTrack has been modified to
store the same.

removeTrack() uses downcasting to remove a TrackPtr from our internal
representation.

The point of invalidateTracksIn() is to remove all tracks belonging to a
certain directory. When a CDS is updated, events are only emitted for
containers (directories). Furthermore, there is no way to know if items
inside were added, deleted or just updated. It is also allowed for
completely new IDs to be generated for the items in the container. To
accomodate all this, invalidateTracksIn() takes a slightly naive route.
It iterates over all known container paths. If /music has been modified,
we have no choice but to invalidate all items (recursively) within it.
So even tracks at /music/Radiohead have to be invalidated. So it simply
checks if any of the paths are *within* the modified path using
startsWith. The alternative would be to maintain a complete tree
mirroring the CDS layout or for each track to have a list of all parents
upto the root of the CDS. Both are even more inefficient/inelegant than
this in my opinion