Select tracks first in the file browser in single-click mode
Previously, a single click on a track will append&play it. This overrides
the behaviour for single-clicks in KDE single-click mode so that it
actually behave more like double-click mode.

Perhaps it is better to decide how single and double-click modes are
handled, what happens when an item is activated, what actions to take
for different types of items, etc. Then it can be implemented inside our
PrettyViews so that all the specialized views can handle mouse clicks more
consistently.

BUG: 233171
CCMAIL: amarok-devel@kde.org