Use KMainWindow::setAutoSaveSettings() for saving window size etc
Previously, the position of the main window is saved on quit and
restored on start up. setAutoSaveSettings() doesn't do this, but this
position should be handled by the window manager anyway. If you want
Amarok to show up at the same screen position each time, that should be
configured in KWin or the WM of your choice.