Added basic UpnpSearchCollection.
The search collection doesn't do any KIO-ing.
It returns a UpnpQueryMaker who can only browse artists,
that too not so well since it overwrites previous entries.
It doesn't return a proper resultCount().
Amarok crashes on exit, most likely due to this.

But hey! it works a bit, even if the design sucks :/