Duration interpretation is now done manually.
The duration format is not strictly conformant to the ISO date format since it is a relative measure, with hours not constrained to 00-23 and so on. In addition the fractions of a second have to be interpreted. So it is done manually. Thanks to Bart Cerneels <bart.cerneels@kde.org> for reporting this