Allow scrobbling option to be changed
fe8cdbf0 introduced a bug whereby the "Submit tracks" checkbox cannot
ever be enabled so that a user can change the setting.  Made it so that
the setting can be changed after the authentication is checked.  An even
better idea would be to check authentication immediately when the dialog
box opens.

CC: unnamedrambler@gmail.com