Got a working two level artist and album focused search query maker.
Turns out that the CollectionTreeItemModelBase enforces that when a sub-tree is expanded, it should finally have atleast one child. This was causing hundreds of query makers to be created leading to the kio slave being overloaded. Now, in the case of no results from the CDS, a fake track is put in just to stop more queries. It probably needs a way to make it really seem invalid.

Sometimes similar queries are launched from different areas of Amarok. A proof of concept query cache is now a part of UpnpQueryMaker, it will be split into a separate class as soon as I figure
out some issues.