wiki: fetch page with the preferred languages
If multiple languages are configured, we first query wikipedia for the
"langlinks" of a given title. Then the first preferred language that
match will be used to fetch the wiki page. If only one language is set,
that is used directly.

The handling of refined search and postfixes have been removed, since
it's not desirable to add those for all languages. However, specific
ones (English, German, etc.) may be added later in order to retain keep
the previous behaviour.