APG: Fix inconsistent translation of constraint names
The constraint names should always be translated in the GUI, and never
translated in the XML file.  Sometimes they were getting translated in
the XML file, which caused problems for users when using localized
versions of Amarok.

And while I was playing around with translation of the constraint names,
I took the opportunity to tweak their appearance in the GUI.

If you're using a localized version of Amarok, please test that the
constraints in APG presets are always exported/imported/loaded/saved
correctly.  Thanks.

BUG: 242111