Fix double click regression in the collection browser
Now double clicking an item will add them to the playlist. This happens for
both single-click and double-click mode in KDE system settings.

Some caveats: Users of double-click mode will need to resort to the
expand/collapse marker for expanding/collapsing; and under single-click
mode, the item will still be expanded/collapsed if it has children.

BUG:244251