Added UpnpQuery.
UpnpQuery abstracts the creation of multiple query URLs to handle boolean ORed queries.
UPnP does not support nested queries and hence handling a 'A and (B or C)' requires
running both 'A and B' and 'A and C' using the distributive property.
This class internally uses a stack and list to simulate a tree. It has currently
been tested only for simple queries