UpnpQueryMaker uses UpnpQuery.
Radical changes in UpnpQueryMaker. It uses the UpnpQuery to build up a query,
makes multiple KIO calls and keeps track that all of them are finished.
It mostly works but fails at resolving nested queries properly, probably because
the server is reporting bad data, probably because of a flaw in the code.
Caching has been switched on again, so that no entries are emitted until
all jobs are done. Doing delta updates would require modification in other
areas of Amarok, which can be done later.

'No results' is no longer emitted since it confuses Amarok which tries to find
tracks/albums etc containing 'No results' as a match/filter.

Filtering can definitely be improved