Always search for audioItem for now.
The rationale has to do with the way QueryMakers are supposed to behave.
Consider an Artist/Album/Track layout. When the search filter bar has the text
'twist', it is expected that every level also searches beneath it. So the QM for
the Artist level has to search for tracks belonging to the artist which have
the pattern, or albums having the pattern.
Since UPnP doesn't support this in any sane manner, we always ask for audioItems
and retrieve the actual data based on query type from the results