albums applet: put tracks as children of disc number item
Tracks that have disc number assigned are created under the disc number
item in the model. This makes the disc number item more useful; it can now
be expanded/collapsed and is draggable by the mouse. When dragging, all
of its children are selected and can be dropped in the playlist, just like
an album item.