Fix MainWindow layout saving if window is hidden
Qt restoreState() and saveState() methods require a valid layouting when
loading/saving the state of the DockWidgets.
On "Minimize to tray" the MainWindow is hidden which destroys this.
Saving the state then gives us an invalid layout.
Same goes for restoring, the state must be restored while the MainWindow
is visible.

BUG: 244583