Added UpnpQueryMakerInternal to the build.
UpnpQueryMakerInternal handles the issue of checking for cache hits and requesting track
meta-data as required, significantly reducing the amount of data transfer and parsing required.
It first checks if the track already exists using listDir. When tracks aren't found, it invokes
individual stat jobs to get information about the tracks. The CACHE_CHECK_THRESHOLD decides
at what fraction (cache size/actual no. of tracks in collection) the caching should be used.

It is a very low possibility that lots of StatJobs will be created since browsing is usually done per
level, and the proper cache fraction will ensure that stat jobs are never launched for a relatively empty cache.