upcoming: put events list in a Plasma::Extender
The plan with using extenders here is for grouping different types of
events. Later the ability to add venue specific events will be
introduced; so a clear way of separating, e.g. "upcoming events for the
current playing artist" and "upcoming events at wherever" is needed.
Plus, extenders support collpasing and dnd rearragement which can be
quite useful.