Fix infinite cpu churning if browser dock is hidden on startup
If the browser dock is put into a tabbed arrangement with some other dock,
and Amarok is started with the browser dock hidden, an endless loop occurs
with the dock trying to resize the collection browser repeatedly. This
happens because the browser is not yet polished, and its delegate tries to
use font sizes in its sizeHint(), as explained in QWidget::ensurePolished()
docs. A simple workaround is to polish it in the dock's constructor.

BUG: 244205