Fix current track applet's handling of playback stopped state
If the context dock is polished after playback starts, it will end up in
the stopped state mode (with last played tracks showing, and also last
added albums in the albums applet).

This fixes that and also removes a timer used for switching states. Now
the current track applet (and hence the albums applet) will be more
responsive on data updates.