APG is never going to implement PlaylistProvider
The problem isn't implementing the APG as a PlaylistProvider.  That's
pretty easy to do.

The problem is implementing APG Presets as a Playlist, which is the
defined behavior of the PlaylistProvider.  The single biggest
showstopper to this implementation is the fact that Playlist::tracks()
is defined synchronously, while the ConstraintSolver in the APG operates
asynchronously.  Thus, either the Playlist class would have to be
substantially rewritten to operare asynchronously, or the call to
Playlist::tracks() would have to block until the ConstraintSolver
finishes.  Neither of these is an attractive option.

Further problems (though not as dire) occur with the addTrack() and
removeTrack() functions, and the trackCount() function.  And having an
Observer of an APG-based Playlist is just pointless because the APG
never adds or removes tracks, unless it generates a completely new
playlist.

So as far as I'm concerned, this issue is dead.  The APG will not ever
implement a PlaylistProvider because it just doesn't fit the design
paradigm.

CCMAIL: bart.cerneels@kde.org