Patch for TagDialog - show Folder when editing multiple Files in the same Directory.
Amarok has the possibility to edit the Tags of Multiple Files e.g. an
Album and shows only the Tags that are exactly the same in all the Files.

The only drawback of this for me was that I didn't see the Location
(sometimes the Foldername, which is often the same for all Files in an
Album, contains some Information I wanted to enter into the Tags) of the
Files and had to check "per Track", remember the Information the
Foldername revealed to me, uncheck "per Track" and enter the Information.

My Patch changes this in a Way, that if all Files are local and in the
same Directory it still displays the Location Label, Location LineEdit &
OpenButton in the TagDialog and sets the Text of the LineEdit to the
Folder where all the Files are.

This way it is possible to see the Folder Name, even when you are
editing multiple Files.

Patch by Stefan Derkits <stefan@derkits.at>. Many thanks!