Better workaround for wrong applet size on startup.
This version also works in the case that Amarok was hidden to tray before
starting, and it does not produce a jerking motion when starting in fullscreen
mode.
Thanks to Felix Geyer <debfx-pkg@fobos.de> for the idea.

It might worth to include this patch in Amarok packages.

CCMAIL: debfx-pkg@fobos.de
CCMAIL: amarok-packagers@googlegroups.com