Switched back to Cagibi powered UPnP Collection
Cagibi is distributed as part of KDE SC 4.5,
and is also available as a external package.

With Amarok 2.4's planned release date in December 2010
it will have to rely on Cagibi rather than Solid.

Although Solid support is available in trunk, it
will not be in any release until KDE SC 4.6 in January
or onwards.