wiki: bring back "refined search" for specific languages
Refined search helps in dealing with artist/album names that are ambiguous.
It did that by appending, for example when using English, "(band)" to the
artist name. Now, this is done by getting a listing of close pages and
similar strings are searched for. This happens after the process involving
langlinks.