Check if a lyrics script is running, and update if one started running later
When resuming playback on startup, and the script manager is created
after the main window and usually hasn't finished loading all the
scripts yet. So when the lyrics applet asks for lyrics it gets rejected.
Now the script manager will announce when it adds a lyrics script so the
applet can respond to it.