Let KNS3 manage install/uninstall of scripts
KNS3 keeps a registry of "stuff" that are installed. If we uninstall things
(i.e. delete them) ourselves the registry will still think they are
installed. Likewise if we install them manually by unpacking then KNS3
won't know about it and the script will not be able to update via KNS3.

Once we depend on kde 4.5 we will have more control over KNS3.