APG: Revert having TagMatch use MetaQueryWidget
Changing TagMatch to use the MetaQueryWidget completely broke handling
of dates in the constraint.  Once I've got the date bugs worked out,
I'll bring back the MetaQueryWidget for TagMatch, since it's a
beneficial change in the long term.  But in the short term, I can't let
2.4.0 go out with a broken APG.

CCMAIL: ralf-engels@gmx.de