Adds new Track Progression mode which only plays queued items
This patch adds a new Track Progression mode which works like "Normal" except that it doesn't keep playing items from the playlist; instead it plays the queued items and then stops.

Motivation:
Most of the time I start playing a track (and maybe queue a few more) but immediately choose "Stop playing after current track". With the mode added in this patch, this last step becomes unnecessary.

Issues:
As an icon (see screenshot) for this mode, I have just reused the icon for the "Normal" mode. Maybe something like the icon for "Edit Queue" would be more appropriate. (This is in: src/playlist/navigators/NavigatorConfigAction.cpp)

Courtesy: Armin Straub