Crash when browsing a category [@ CollectionTreeItem::row]
patch for https://bugs.kde.org/show_bug.cgi?id=233170

When a query is started the Collections::QueryMaker* is stored in QHash<CollectionTreeItem*, Collections::QueryMaker*> runningQueries.

Now, when CollectionTreeItemModelBase::populateChildren deletes items CollectionTreeItemModelBase::itemAboutToBeDeleted is called, which in turn stops, disconnects and deletes the queryMaker associated with the item.

The problem occurs, if there is more than one queryMaker for the item, as will happen, when CollectionTreeItemModelBase::handleCompilations is called. In this case itemAboutToBeDeleted deletes only one queryMaker. The other one finishes normally, calls CollectionTreeItemModelBase::queryDone, which in turn calls CollectionTreeItem::row of a deleted item and Amarok crashes with a segfault.

The patch changes the runningQueries buffer to QMultiHash. So all queryMakers are kept and itemAboutToBeDeleted can stop them all.

Courtesy: Wolfgang Lorenz