More intelligent automatic lyrics scrolling.
Now it is possible for a user to scroll in the lyrics, even if
auto scrolling is enabled. The auto scrolling remembers the user
set offset and takes it into account.
This way the user can correct errors in the position guessing, which
will fail e.g. for songs with long intros/outros.