Collapsed 4 actions in the Playlist toolbar into 1 menu.
With the recent move of the tracks count from the status bar to the
playlist toolbar, and the disposal of the status bar, the playlist
toolbar has just grown too wide, and is not nicely usable on a netbook.
This commit moves 4 of the playlist actions into a menu in the playlist
toolbar, so we gain the horizontal space equivalent of 3 icons.
I dislike useless clicking as much as anybody so this solution is not
ideal, but not being able to make the playlist narrower is too much of
a deal breaker on netbooks.
Also, the icons are somewhat vertically misaligned by a pixel or two,
probably because they were done in different moments. We might consider
having them fixed.