iPod handler: Correctly read & save last track played time from/to iPods
libgpod returns zero last played time to signify "never played",
interpret that as invalid QDateTime as in rest of Amarok.

Also implement write functionality where exactly inverse transformation
is performed.

Big thanks to Christophe Fergeau for noticing suspicious workaround that
turned out to be a sign of an error in different part of the patch.